Initial spike.
Add an initial spike of what a Behaviour looks like--a description, a
function to call to generate the HTTP request to make, and a function to
call to check the returned response.

Add a Runner interface (that can be filled by the default HTTP client)
and a Run function that loops through behaviors, running the requests
they generate using the passed Runner, checking their responses, and
using the responses as the input for the next request generator. This
leads to a test that operates something like this:

1. Generate request using Behaviour 1's RequestFunc (no response arg)
2. Run the request with Runner
3. Check the response from the Runner using Behavior 1's CheckFunc (args
are return parameters from step 2)
4. If the check fails, return the error. If the check passes, continue.
5. Generate request using Behaviour 2's RequestFunc (response arg from
step 2)
6. Run, check, etc.

This is repeated until we're out of behaviours, and is useful in
building a user interaction flow against an API that may have
server-defined outputs that are needed in future inputs.